requirements.txt
os
dotenv
telebot
pinecone-client
langchain